Split knit model.rs into per-domain modules
Break the 900-line src/model.rs into config (workspace config + context map),
project (templates: repos, runtime, landing), org, workitem, and bundle
(ChangeGroup + repos/commit groups/ledger nodes/publications). model/mod.rs
re-exports everything and holds the shared SCHEMA_VERSION/checkout-mode
constants, so all `crate::model::<Type>` call sites are unchanged. No behavior
change; tests and clippy unchanged.


Knit-Group: kg_20260531_08eaf3
Knit-Bundle: split-oversized-files-round-4
